TODO before pushing to master: (1) Refactor colRangeToIndices from current index-based solution (which                                                                   is based                                                                   on the                                                                   way                                                                   it's                                                                   done                                                                   for                                                                   ranges) to a cell-based solution. Currently I take the cells in the colRange, convert to a range, the range converts to indices, the indices convert back to cells. This appears to work but is inefficient.
(2)  Move column range helper methods into the appropriate files.
(Either DB/Eval for the initial ancestor lookup, or part of Eval/Core
 for interpolating from EvalContext.)

(3) Leave clear comments indicating that column ranges ancestors are
*not* all loaded into the EvalContext at getAncestors, but rather during
list interpolation. Enough information exists in the EvalContext to do
list interpolation properly, but leave a very clear comment regarding
what the EvalContext can be when getAncestors is called on a colRange.

(4) Refactor the method I replaced repeatedly, (contextMap ctx), which didn't work
for colRanges given the EvalContext is not complete at time of ancestor
loading. Create a funtion modifiedLookupInContextMap ::
EvalContext -> (M.map ASIndex ASCell) to do the same thing as the
workaround I replaced (contextMap) ctx with.

(5) Make A:A a valid column range (right now, only A1:A or A2:A works).

(6) Write backend tests for relative and absolute
references, undo, decoupling, etc. This is important after I refactor
the colRange to not use the same functions range is using.

(7) Make sure the relative references shift appropriately when changing cells.
(E.g. when A1:$B is in the spreadsheet, and A is shited to D, then the
 remaining range should be $B1:D).

(8) Add rows.