evaluateFormat kernel endpoint created for conditional formatting
Summary:
Python kernel has a "run_raw" API endpoint that lets you run code against a sheet
without any serialization/AST transformation. This corresponds to a "evaluateFormat" function
in Kernels/Python that returns a FormatResult. See Types/CondFormat for the definition of this
type.

TODO:
Alex needs to fill in the parseFormatResult parser in Parsing/Read, and tie it to frontend.

Test Plan: Test with frontend once the above TODO is complete.

Reviewers: zhukeepa

Differential Revision: http://18.102.225.27:11000/D12