Build backend on Mac.
Summary:
I really want it to be possible for everyone to build the backend. It just
seems like an organizational health thing.

I'm not sure how important the docker thing is, but I removed it for now since
it was my main blocker to building with stack.

I also added some instructions for installing R dependences.

I also upgraded `inline-r` to avoid a breaking
[issue](https://github.com/tweag/HaskellR/issues/220).

Test Plan:
$ stack install
$ stack exec alphasheets-exe

Reviewers: anand

Reviewed By: anand

Subscribers: michael, timchu

Differential Revision: http://18.102.225.27:11000/D19