Fix javascript error in tooltips.
Summary:
The issue was that react-tooltip would hold a reference to the tooltip
node (the one with `data-tip` and `data-for`), which it expected to
always be there. We were adding and removing it from the dom though.

In the end the fix was to patch react-tooltip. Will also send a pull
request so hopefully this gets merged in to the original.

Most of this is just formatting, but since I did it already, we might as well
use it.

Test Plan: (see attached video)

Reviewers: RITESH

Reviewed By: RITESH

Differential Revision: http://18.102.225.27:11000/D40