fix lambda formatting
Summary:
The Format class's __repr__ was wrong -- it gave back a value that Haskell couldn't actually parse. Fixed that.

(gulp test-eval had actually been failing. The reason this went unnoticed was that the Python kernel server did *not* reboot when we updated the Python code, and all the tests passed with the old code.)

Test Plan: gulp test-eval should *actually* pass. pull it locally and run the tests

Reviewers: anand, timchu, RITESH

Reviewed By: RITESH

Differential Revision: http://18.102.225.27:11000/D32