Fix decoupling and height of textbox.
Summary: Apparently "px" *does* matter in this case.

Test Plan:
Open `localhost:8080`, double click in A1, see that the textbox that
appears is approximately the same size as the box. Input (Python)
`range(10)`, see that the first 10 rows are filled with 0..9. Click back
in A1, hit delete, get an alert asking if I'm sure I want to decouple.

Reviewers: zhukeepa

Reviewed By: zhukeepa

Differential Revision: http://18.102.225.27:11000/D33