fix .gitignore to actually include /static/images/dummy while excluding everything we don't want
Summary: We previously had a number of problems where, upon pulling from a remote branch, the /static/images folder was missing. The reason for this was a screwy .gitignore. This commit *actually* fixes it.

Test Plan:
Create a file dummy2 inside of /static/images and a file dummy3 inside of /static. Modify the dummy file. Run:

git rm --cached . -r -f ; git add -A; git status

The *only* file staged for the change should be /static/images/dummy.

Reviewers: anand, RITESH, timchu, michael, joel

Reviewed By: joel

Differential Revision: http://18.102.225.27:11000/D36