fix textbox location
Summary:
Reverted the textbox position logic to the old version, which is much uglier and less react, but doesn't make the textbox appear in the incorrect location.

The problem before is that ASSpreadsheet doesn't rerender when Textbox rerenders, so the position prop doesn't get updated, even when the
selection does. This is kind of ugly, but I'd rather not try to do this "correctly" when ASSpreadsheet is already so fucked -- seems
like a deep rabbithole that I don't yet want to crawl into.

Test Plan: it should basically work as it did before

Reviewers: anand, RITESH, michael

Reviewed By: michael

Differential Revision: http://18.102.225.27:11000/D52