ctrl+enter should now always eval
Summary: change it to eval only if expression is not empty, and cell is not part of an expanding cell

Test Plan:
TEST 1
1) Eval range(10)
2) Press ctrl+enter in one of the cells in range(10).
You should NOT be asked to decouple.

TEST 2
1) Eval a=1 in the header.
2) Eval a in A1.
3) Eval a=2 in the header.
4) Eval a in A1.

A1 should update to 2.

TEST 3
1) Press Ctrl+Enter in a blank cell.
You should not see any message to the server.

TEST 4
1) Eval range(10)
2) Eval range(15) in A1.
You should see 0-14 in A1-A15.

Reviewers: RITESH, michael, anand, zhukeepa

Reviewed By: anand, zhukeepa

Differential Revision: http://18.102.225.27:11000/D51