Fixed undo on setProps.
Summary:
Fixed undo on setProps. We were extraneously setting cells in the DB and screwing up the guarantee that, when a commit is pushed in dispatchcycle, the previous items in the sheet were the items in the DB. In the future, we shouldn't set cells in the DB in any handler that calls runDispatchCycle.
Added a decently comprehensive test for undo and redo on setProps.

Test Plan: run gulp test-eval.

Reviewers: RITESH, zhukeepa

Reviewed By: zhukeepa

Subscribers: RITESH, zhukeepa

Differential Revision: http://18.102.225.27:11000/D61