Remove ASCellStore setters.
Summary:
This is setup to making ASCellStore a ReduceStore. First step is
removing this unprincipled mutation.

Test Plan: Key and click around to random cells in the grid. Use the input to enter 123 in the A1. Click on A2, see the editor clear. Arrow between A1 and A2, seeing the editor take on the right value each time. Click between the two, seeing the same. Enter 234 in A2 using the editor, repeat toggling behavior.

Reviewers: michael, zhukeepa, anand

Reviewed By: anand

Differential Revision: https://alphasheets.phacility.com/D99