Databinding refactor
Summary:
The databinding refactor.

(1) CodeEditor and Textbox mirror state through ExpressionStore.
(2) The onKeyDown of each component (grid, editor, textbox) are all really simple and execute either Expression actions or Shortcut actions
(3) Shortcuts are actions, because there's no possibility of directly mutating editors now.
(4) Event handlers in evalPane that needed to be there due to ref mutation are moved to App.
(5) The focusable HOC keeps the app focus synchronized with FocusStore
(6) SelectionStore is immutable
(7) A ConfigurationStore manages the view configuration (header open, bottom pane, etc.). This is necessary for (a) changing view config from shortcuts, and (b) autofocusing views when they materialize
(8) Spreadsheet is selection-controlled, thus its selection can be trivially changed with an action anywher
(9) There are a bunch of additions to Ace that this required (special event listeners). They're on my fork.

There are a lot more details, but in general, you can now do basically anything with an action in AlphaSheets.

Test Plan: TODO. I've taken it through a bunch of the manual tests, some of them are failing due to stupid reasons. It's not blocking this review, though.There are some details to resolve, like ref insertion and autoscrolling, but at this point they're details.

Reviewers: michael, RITESH

Subscribers: zhukeepa

Differential Revision: https://alphasheets.phacility.com/D144