auto-focus to header pane on eval, if output present
Summary:
* small refactors in various parts of code with listeners
* split HeaderStore into HeaderStore and HeaderOutputStore
* If the header is evaluated AND there is output, the header pane is brought to focus.

Test Plan:
1. Set the language to Python.
2. Click on Code -> Toggle Header.
3. Type a=5.
4. Press Ctrl+S.
You should get a message just saying that the header saved; nothing else should happen.
5. Below a=5, type a+1, and evaluate. (The header should be two lines now:
a=5
a+1)
You should see a popup with the number 6 in it.
6. Click the leftmost button on the bottom.
You should see a blank popup at the bottom.
7. Below a+1, add the line "print a", so the header should now be three lines:
a=5
a+1
print a
8. Press evaluate.
You should see a popup that says "None" a bunch of hyphens, then 5.
9. Click the leftmost button again.
10. Press evaluate again.
You should be brought back to the popup.

Reviewers: anand, RITESH, michael

Differential Revision: https://alphasheets.phacility.com/D154