error and cell output change color depending on selection
Summary: now if an active cell has an error, or output content, turn the bottom icons orange

Test Plan:
1. Set the language to excel, and type =ASDF into A1.
2. Set the language to Python, and type a32es into B1.
3. Click on A1.
The bottom left icon should be orange, the middle icon should be white.
4. Press right.
Both the bottom left icon and middle icon should be orange.
5. Enter "print 15" and press Ctrl+Enter.
The active cell should be B1, the middle icon should be orange, and the bottom left one should not be.

Reviewers: anand, joel, RITESH

Reviewed By: RITESH

Differential Revision: https://alphasheets.phacility.com/D168