Reduce duplication and solve half of an overlay bug
Summary: We refactored ASOverlayController to accept the same getPixelCoordinates function as a prop that Textbox accepts, and so unified the logic. The old (separate) function for getting pixel coordinates in ASSpreadsheet for ASOverlayController was also incorrect, and the Textbox version was correct, so this eliminated some bugs (described in test plan). However, Hypergrid still has a bug that prevents this implementation from fully working (the test case that does not pass is also described in the test plan). The bug in Hypergrid is that getBoundsOfCell sometimes returns a value that would've been correct for the previous value of getScroll, even though the user has scrolled since then. (i.e. getBoundsOfCell is returning values that aren't always consistent with the rest of state)

Test Plan:
Test that passes, that didn't pass before this diff

```
(1) Plot an image in C1
(2) Resize the column C to be bigger than the other columns, while keeping all the other columns 100px.
(3) Scroll right until the viewing window shifts.
(4) Click away.
(5) The position of the plot overlay should change upon click to an erroneous location (off by 140px).
```

Test that does not pass

```
(1) Plot an image in C1
(2) Scroll right until the viewing window shifts.
(3) Scroll left until you hit column B.
(4) Scroll left one more time, until the viewing window shifts to include column A again.
(5) The moment after you scroll left, the position of the plot overlay should be correct (and not off by 140px).
(Note: with this diff, if you click away after step (5), the position will be correct upon rerender.)
```

Reviewers: RITESH, anand

Reviewed By: anand

Differential Revision: https://alphasheets.phacility.com/D209