Make sure that the line beginning with = is always visible in the code editor.
Test Plan:
Let cell A1 have the expression
```
def f(x):
  return x + 1

=f(4)
```

Then:

(1) If the expression is typed in the textbox, not the code editor, then the moment we type =, the code editor should jump to line 4.
(2) The moment we backspace the = while focused on the textbox, the code editor should go back to line 1.
(2) In any case, after evaluating the cell, selecting the cell again should cause the code editor to jump to line 4.

Reviewers: anand, RITESH, zhukeepa

Reviewed By: zhukeepa

Differential Revision: https://alphasheets.phacility.com/D213