Fix the blank Hypergrid problem by forcing a repaint on every window focus. Also, fix flow.
Summary: On Macs with dual GPUs, the canvas that supports Hypergrid occasionally goes blank. When it receives a repaint request, it works again (consistently). Therefore we should repaint on every window focus, so that the user never comes back to a blank page (a common experience for me).

Test Plan: Empirically, the page has stopped being blank sometimes

Reviewers: RITESH

Reviewed By: RITESH

Differential Revision: https://alphasheets.phacility.com/D214