Forcibly reflow spreadsheet on window.onfocus
Summary: On Macs (especially with dual GPUs), there is a recurring but irreproducible problem with losing graphics context and thereby making the canvas of hypergrid blank. We fix this by setting display: 'none' and then display: 'block' 250ms after, whenever the window is focused, since the problem only ever occurs when we switch back into the AlphaSheets tab after being on another tab for a while.

Test Plan: Empirically, AlphaSheets should no longer be blank when switching back to the window. Sample size n=10, this seems to have happened on my computer.

Reviewers: RITESH, anand, zhukeepa

Reviewed By: zhukeepa

Differential Revision: https://alphasheets.phacility.com/D217