Cross-sheet references & cleaner argument passing through handlers
Summary:
Cross-sheet refs work. This would have added another two arguments (userId and db connection) to a bunch of functions cascading down from the handlers, so instead of adding to the cruft I cleaned it up with a MessageContext datatype. This should make it pretty easy to put our functions into (ReaderT MessageContext IO a) whenever we decide to do that.
Notice that this isn't the naive implementation where user sheets are queried on every excel reference lookup; sheet lookups are batched upon every call to eval.

Test Plan:
- '1' in A1
- new sheet 'Sheet2'
- '=Sheet1!A1 * 10' in A1, A1 is 10
- open sheet 'Sheet1'
- change A1  to 100
- open 'Sheet2' again, A1 is now 100

Reviewers: sdiehl, RITESH, zhukeepa

Reviewed By: zhukeepa

Differential Revision: https://alphasheets.phacility.com/D253