Fix CSS problem on Macs
Summary: Fix CSS problem on Macs, flex some things, get rid of hardcoded 85% width, 750px width, make bottom bar look not shitty

Test Plan:
On a Mac, open AlphaSheets and:
(1) Before, the tabs were not clickable. After this diff, you can click on another tab and switch to it.
(2) Before, the SheetBrowser was hardcoded to be 85% of the total width of the window, and the tabs section was set to 750px. After this diff, there should be a constant spacing between the tabs and the buttons on the right of the bottom bar.
(3) Before, the tabs had a hard edge. After this diff, the tabs "blend in" to the background.

Reviewers: joel, RITESH, anand

Reviewed By: anand

Differential Revision: https://alphasheets.phacility.com/D262